requirements.txt
# Web Framework
fastapi>=0.68.0
uvicorn>=0.15.0

# Database
sqlalchemy>=1.4.0
mysql-connector-python>=8.0.0
pymysql>=1.0.0
cryptography>=3.4.0
alembic>=1.7.0

# Cache & Message Queue
redis>=4.0.0
aioredis>=2.0.0

# Data Processing
pandas>=1.3.0
numpy>=1.21.0

# API Client
httpx>=0.23.0
websockets>=10.0

# Security
python-jose>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.0,<5.0.0
python-multipart>=0.0.5

# Environment Variables
python-dotenv>=0.19.0

# Validation
pydantic>=1.8.0
email-validator>=1.1.0

# Monitoring
prometheus-client>=0.11.0

# Testing
pytest>=6.2.0
pytest-asyncio>=0.15.0
pytest-cov>=2.12.0

# Development
black>=21.7b0
flake8>=3.9.0
mypy>=0.910
isort>=5.9.0
